requirements.txt
opencv-python 
mss 
numpy 
pyautogui
PyYAML